feat: run start task 不执行删除构建目录

+++ src/tasks/build.ts
@@ -4,6 +4,7 @@ import { buildLess } from './less';
 import { copyAssets } from './assets';
 import { buildTs } from './typescript';
 import { getDirPath } from '../utils/utils';
+import debug from '../debug';
 
 const esDir = getDirPath('es');
 const libDir = getDirPath('lib');
@@ -11,8 +12,13 @@ const libDir = getDirPath('lib');
 export function build(modules?: boolean) {
   const dir = modules === false ? esDir : libDir;
 
-  // 删除输出目录
-  rimraf.sync(dir);
+  const taskName = process.env.TASK_NAME;
+
+  if (taskName !== 'start') {
+    debug.log(`rm build ${modules === false ? 'es' : 'lib'} dir`);
+    // 删除输出目录
+    rimraf.sync(dir);
+  }
 
   return merge2([
     buildLess(dir),
